changed the structure of the files to do with the image components
