--- src/components/buildingImages.js
@@ -1,52 +1,0 @@
-import Bank from "../assets/images/Bank.png"
-import Bank_locked from "../assets/images/Banklocked.png"
-import Bridge from "../assets/images/Bridge.png"
-import Bridge_locked from "../assets/images/Bridgelocked.png"
-import City_scape from "../assets/images/City_scape.png"
-import Empty_city from "../assets/images/Empty_city.png"
-import Fire_station_ from "../assets/images/Fire_station_.png"
-import Fire_station_locked from "../assets/images/Fire_station_locked.png"
-import Hospital from "../assets/images/Hospital_.png"
-import Hospital_locked from "../assets/images/Hospital_locked.png"
-import Houses from "../assets/images/Houses.png"
-import Houses_locked from "../assets/images/Houseslocked.png"
-import Museum from "../assets/images/Museum.png"
-import Museum_locked from "../assets/images/Museumlocked.png"
-import Park from "../assets/images/Park.png"
-import Park_locked from "../assets/images/Parklocked.png"
-import Recycle_Plant from "../assets/images/Recycle_Plant.png"
-import Recycle_Plant_locked from "../assets/images/Recycle_Plantlocked.png"
-import Town_Hall from "../assets/images/Town_Hall.png"
-import Town_hall_locked from "../assets/images/Town_Halllocked.png"
-import Train_station_ from "../assets/images/Train_station_.png"
-import Train_station_locked from "../assets/images/Train_station_locked.png"
-
-
-const images = {
-    Bank,
-    Empty_city,
-    Bridge,
-    Fire_station_,
-    Hospital,
-    Houses,
-    Museum,
-    Park,
-    Recycle_Plant,
-    Town_Hall,
-    Train_station_
-}
-
-const imagesLocked = {
-    Bank: Bank_locked,
-    Bridge: Bridge_locked,
-    Fire_station_: Fire_station_locked,
-    Hospital: Hospital_locked,
-    Houses: Houses_locked,
-    Museum: Museum_locked,
-    Park: Park_locked,
-    Recycle_Plant: Recycle_Plant_locked,
-    Town_Hall: Town_hall_locked,
-    Train_station_: Train_station_locked
-}
-
-export { images, imagesLocked }
